# front/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdoptMe API Documentation",
    "description": "Especificación interactiva de los endpoints de la plataforma de adopción de mascotas AdoptMe. Incluye gestión de usuarios, roles, catálogo de mascotas y solicitudes de adopción.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo local"
    }
  ],
  "paths": {
    "/api/auth": {
      "post": {
        "summary": "Iniciar sesión (Login)",
        "description": "Autentica a un usuario mediante correo electrónico, contraseña y rol esperado.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@test.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["usuario", "empresa", "administrador"],
                    "example": "usuario"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Login exitoso" },
                    "usuario": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "nombre": { "type": "string", "example": "Juan Pérez" },
                        "email": { "type": "string", "format": "email" },
                        "rol": { "type": "string", "example": "usuario" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos faltantes en la solicitud"
          },
          "401": {
            "description": "Credenciales inválidas o conflicto de rol"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Gestión de recuperación de contraseña",
        "description": "Maneja el flujo de 'Olvidé mi contraseña' (solicitud de código, verificación de código y restablecimiento de contraseña).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "juanes15342@gmail.com"
                  },
                  "action": {
                    "type": "string",
                    "enum": ["request", "verify", "reset"],
                    "example": "request"
                  },
                  "code": {
                    "type": "string",
                    "description": "Código de 6 dígitos (Requerido para 'verify' y 'reset')",
                    "example": "123456"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Nueva contraseña (Requerido para 'reset')",
                    "example": "nuevapassword123"
                  }
                },
                "required": ["email", "action"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Código de recuperación enviado" },
                    "devCode": { "type": "string", "description": "Código generado (Disponible solo en desarrollo)", "example": "123456" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Código incorrecto, expirado o contraseña no cumple con los requisitos de longitud"
          },
          "404": {
            "description": "Correo electrónico no registrado"
          }
        }
      }
    },
...
  },
...
}